Updated with Mark Calabretta's changes:
Replace the variadic macro, WCSERR_SET, with one that fills in the
first portion of the argument list.  The main reason was that C89
doesn't have variadic macros.  However, the new version is more suited
to creating macros that are tailored to the particular source file,
e.g.  WCS_ERRMSG in wcs.c.  I also changed the wcserr struct so that
it records the name of the function where the error occurred, and
modified the wcserr_set and its various helper macros accordingly.

Apart from the Fortran wrappers and the documentation, the other
significant change was to replace the wcserr struct in wcsprm and the
various other structs with just a pointer.

In wcs.c, I've changed things a bit to avoid using wcserr_copy() as
much as possible.  The full hierarchy of error messages is preserved
and printed by a new function, wcsperr(), which produces a traceback-
type error listing.  Also, rather than save errors up and set them on
exit from a function, particularly in the prj.c functions, I've mainly
called wcserr_set() at the point where the error occurs so that the
line number info will be more meaningful.